+++ libs/pavexc/src/compiler/analyses/user_components/router.rs
